can drag and drop .obj files and download .sprig.obj
